requirements.txt
streamlit>=1.41.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.22.0
opencv-python-headless>=4.8.0
Pillow>=9.5.0
moviepy>=1.0.3
imageio>=2.31.1
imageio-ffmpeg>=0.4.8
scikit-image>=0.21.0
matplotlib>=3.7.1
tqdm>=4.65.0
face-alignment>=1.3.3
PyYAML>=6.0.1
scipy>=1.10.1
pandas>=2.0.3